+++ src/main/java/insane96mcp/iguanatweaksexpanded/mixin/CrossbowItemMixin.java
@@ -1,15 +1,13 @@
 package insane96mcp.iguanatweaksexpanded.mixin;
 
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.entity.LivingEntity;
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import net.minecraft.world.item.CrossbowItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ProjectileWeaponItem;
-import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(CrossbowItem.class)
 public abstract class CrossbowItemMixin extends ProjectileWeaponItem {
@@ -17,9 +15,10 @@ public abstract class CrossbowItemMixin extends ProjectileWeaponItem {
         super(pProperties);
     }
 
-    @Inject(method = "shootProjectile", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/Level;addFreshEntity(Lnet/minecraft/world/entity/Entity;)Z", shift = At.Shift.BEFORE))
-    private static void onShootProjectile(Level pLevel, LivingEntity pShooter, InteractionHand pHand, ItemStack pCrossbowStack, ItemStack pAmmoStack, float pSoundPitch, boolean pIsCreativeMode, float pVelocity, float pInaccuracy, float pProjectileAngle, CallbackInfo ci) {
-
+    @Inject(method = "getChargeDuration", at = @At(value = "RETURN"), cancellable = true)
+    private static void onGetChargeDuration(ItemStack stack, CallbackInfoReturnable<Integer> cir) {
+        if (stack.getItem() instanceof CrossbowItem crossbowItem && stack.getEnchantmentLevel(NewEnchantmentsFeature.CURSE_OF_SLOW_CHARGE.get()) > 0) {
+            cir.setReturnValue(cir.getReturnValue() + 10);
+        }
     }
-
 }