+++ changelog.md
@@ -5,6 +5,8 @@
   * A bow enchantment that makes arrows ignore gravity for a while
 * Added Sprint pact
   * A new pants treasure enchantment that greatly increases movement speed at the cost of not being able to run anymore
+* Added Curse of Slow charge
+  * Opposite of Quick Charge
 * Reworked M.A. Jump
   * Now called Double Jump, it's one level only and is a treasure enchantment
   * Any equipment with it should be disenchanted as it has an invalid enchantment now